Repeat statements with u32 parameters

+++ assembly/src/parsers/nodes.rs
@@ -9,7 +9,7 @@ use core::fmt;
 pub enum Node {
     Instruction(Instruction),
     IfElse(Vec<Node>, Vec<Node>),
-    Repeat(u16, Vec<Node>),
+    Repeat(u32, Vec<Node>),
     While(Vec<Node>),
 }
